Consolidate PyBlockStore impls
This moves the PyBlockStore BlockStore impl to where the PyBlockStore
ChainReader impl is, reduces to one structs named PyBlockStore, and
implements BatchIndex, and TransactionIndex for PyBlockStore.
